+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 
  - Fix pac `defmt` feature
+ - Fix timer interrupt status clear
 
 ## [v0.22.0] - 2024-10-04